Move non-vulnerable namespace and Thing to the sequence view
