--- testing/vulnerable_app/models/thing.py
@@ -1,8 +1,0 @@
-from ..core.database import Base
-
-
-class Thing(Base):
-    """
-    This is separate from Widget since reusing it causes flakiness.
-    """
-    pass
